+++ js/main.js
@@ -1,6 +1,5 @@
 import {activatorFormAndFilters, inactiveFormAndFilters} from './work-with-form.js';
-import {createMap} from './map.js';
-import {createSlider} from './slider.js';
+import {createMap, createSlider} from './map.js';
 inactiveFormAndFilters();
 createMap();
 activatorFormAndFilters();
